chinh sua giao dien all page + update add category
