+++ .idea/artifacts/CabbageNews_war_exploded.xml
@@ -15,6 +15,49 @@
           <element id="library" level="project" name="Maven: com.googlecode.objectify:objectify:5.1.22" />
           <element id="library" level="project" name="Maven: javax.servlet:javax.servlet-api:3.1.0" />
           <element id="library" level="project" name="Maven: jstl:jstl:1.2" />
+          <element id="library" level="project" name="Maven: com.fasterxml.jackson.core:jackson-core:2.9.6" />
+          <element id="library" level="project" name="Maven: com.google.api.grpc:grpc-google-cloud-pubsub-v1:1.56.0" />
+          <element id="library" level="project" name="Maven: com.google.api.grpc:proto-google-cloud-pubsub-v1:1.56.0" />
+          <element id="library" level="project" name="Maven: com.google.api.grpc:proto-google-common-protos:1.15.0" />
+          <element id="library" level="project" name="Maven: com.google.api.grpc:proto-google-iam-v1:0.12.0" />
+          <element id="library" level="project" name="Maven: com.google.api:api-common:1.8.1" />
+          <element id="library" level="project" name="Maven: com.google.api:gax-grpc:1.44.0" />
+          <element id="library" level="project" name="Maven: com.google.api:gax:1.44.0" />
+          <element id="library" level="project" name="Maven: com.google.auth:google-auth-library-credentials:0.15.0" />
+          <element id="library" level="project" name="Maven: com.google.auth:google-auth-library-oauth2-http:0.15.0" />
+          <element id="library" level="project" name="Maven: com.google.cloud:google-cloud-core-grpc:1.74.0" />
+          <element id="library" level="project" name="Maven: com.google.cloud:google-cloud-core:1.74.0" />
+          <element id="library" level="project" name="Maven: com.google.cloud:google-cloud-pubsub:1.74.0" />
+          <element id="library" level="project" name="Maven: com.google.code.findbugs:jsr305:3.0.2" />
+          <element id="library" level="project" name="Maven: com.google.errorprone:error_prone_annotations:2.3.2" />
+          <element id="library" level="project" name="Maven: com.google.http-client:google-http-client-jackson2:1.29.0" />
+          <element id="library" level="project" name="Maven: com.google.http-client:google-http-client:1.29.1" />
+          <element id="library" level="project" name="Maven: com.google.j2objc:j2objc-annotations:1.1" />
+          <element id="library" level="project" name="Maven: com.google.protobuf:protobuf-java-util:3.7.1" />
+          <element id="library" level="project" name="Maven: com.google.protobuf:protobuf-java:3.7.1" />
+          <element id="library" level="project" name="Maven: commons-codec:commons-codec:1.10" />
+          <element id="library" level="project" name="Maven: commons-logging:commons-logging:1.2" />
+          <element id="library" level="project" name="Maven: io.grpc:grpc-alts:1.19.0" />
+          <element id="library" level="project" name="Maven: io.grpc:grpc-auth:1.19.0" />
+          <element id="library" level="project" name="Maven: io.grpc:grpc-context:1.19.0" />
+          <element id="library" level="project" name="Maven: io.grpc:grpc-core:1.19.0" />
+          <element id="library" level="project" name="Maven: io.grpc:grpc-grpclb:1.19.0" />
+          <element id="library" level="project" name="Maven: io.grpc:grpc-netty-shaded:1.19.0" />
+          <element id="library" level="project" name="Maven: io.grpc:grpc-protobuf-lite:1.19.0" />
+          <element id="library" level="project" name="Maven: io.grpc:grpc-protobuf:1.19.0" />
+          <element id="library" level="project" name="Maven: io.grpc:grpc-stub:1.19.0" />
+          <element id="library" level="project" name="Maven: io.opencensus:opencensus-api:0.19.2" />
+          <element id="library" level="project" name="Maven: io.opencensus:opencensus-contrib-grpc-metrics:0.19.2" />
+          <element id="library" level="project" name="Maven: io.opencensus:opencensus-contrib-http-util:0.19.2" />
+          <element id="library" level="project" name="Maven: javax.annotation:javax.annotation-api:1.3.2" />
+          <element id="library" level="project" name="Maven: javax.validation:validation-api:2.0.1.Final" />
+          <element id="library" level="project" name="Maven: org.apache.commons:commons-lang3:3.5" />
+          <element id="library" level="project" name="Maven: org.apache.httpcomponents:httpclient:4.5.5" />
+          <element id="library" level="project" name="Maven: org.apache.httpcomponents:httpcore:4.4.9" />
+          <element id="library" level="project" name="Maven: org.codehaus.mojo:animal-sniffer-annotations:1.17" />
+          <element id="library" level="project" name="Maven: org.json:json:20180813" />
+          <element id="library" level="project" name="Maven: org.jsoup:jsoup:1.11.3" />
+          <element id="library" level="project" name="Maven: org.threeten:threetenbp:1.3.3" />
         </element>
       </element>
     </root>
