--- .idea/libraries/Maven__com_google_code_findbugs_jsr305_3_0_2.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Maven: com.google.code.findbugs:jsr305:3.0.2">
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/com/google/code/findbugs/jsr305/3.0.2/jsr305-3.0.2.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/com/google/code/findbugs/jsr305/3.0.2/jsr305-3.0.2-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/com/google/code/findbugs/jsr305/3.0.2/jsr305-3.0.2-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
