+++ .idea/libraries/Maven__org_apache_maven_shared_maven_shared_utils_0_3.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Maven: org.apache.maven.shared:maven-shared-utils:0.3">
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/apache/maven/shared/maven-shared-utils/0.3/maven-shared-utils-0.3.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$MAVEN_REPOSITORY$/org/apache/maven/shared/maven-shared-utils/0.3/maven-shared-utils-0.3-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/apache/maven/shared/maven-shared-utils/0.3/maven-shared-utils-0.3-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
