+++ src/main/java/com/cabbage/servlet/admin/AddArticleServlet.java
@@ -13,7 +13,7 @@ import static com.googlecode.objectify.ObjectifyService.ofy;
 public class AddArticleServlet extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        req.getRequestDispatcher("/admin/article/category-article-form.jsp").forward(req, resp);
+        req.getRequestDispatcher("/admin/article/category-article-crawler-source-form.jsp").forward(req, resp);
     }
 
     @Override
